[CSL-2018] Add migration roundtrip specs
