+++ wallet-new/test/MarshallingSpec.hs
@@ -3,11 +3,13 @@ module MarshallingSpec where
 import           Universum
 
 import           Cardano.Wallet.API.V1.Errors (WalletError)
+import           Cardano.Wallet.API.V1.Migration.Types (Migrate (..))
 import           Cardano.Wallet.API.V1.Types
 import           Cardano.Wallet.Orphans ()
 import           Data.Aeson
 import           Data.Typeable (typeRep)
 import           Pos.Crypto (PassPhrase, emptyPassphrase)
+import qualified Pos.Wallet.Web.ClientTypes.Types as V0
 import           Test.Hspec
 import           Test.Hspec.QuickCheck
 import           Test.QuickCheck
@@ -21,6 +23,7 @@ import qualified Pos.Core as Core
 spec :: Spec
 spec = describe "Marshalling & Unmarshalling" $ do
     describe "Roundtrips" $ do
+        -- Aeson roundrips
         aesonRoundtripProp @BackupPhrase Proxy
         aesonRoundtripProp @AssuranceLevel Proxy
         aesonRoundtripProp @Payment Proxy
@@ -40,6 +43,11 @@ spec = describe "Marshalling & Unmarshalling" $ do
         aesonRoundtripProp @NodeInfo Proxy
         aesonRoundtripProp @NodeSettings Proxy
 
+        -- Migrate roundrips
+        migrateRoundtripProp @AssuranceLevel @V0.CWalletAssurance Proxy Proxy
+        migrateRoundtripProp @WalletId @(V0.CId V0.Wal) Proxy Proxy
+        migrateRoundtripProp @(WalletId, AccountId) @V0.AccountId Proxy Proxy
+
     describe "Invariants" $ do
         describe "password" $ do
             it "empty string decodes to empty password" $
@@ -51,10 +59,21 @@ spec = describe "Marshalling & Unmarshalling" $ do
                 -- currently passphrase should be either empty or of length 32
                 decodingFails @PassPhrase "aabbcc" Proxy
 
+migrateRoundtrip :: (Arbitrary a, Migrate a b, Migrate b a, Eq a, Show a) => proxy a -> proxy b -> Property
+migrateRoundtrip (_ :: proxy a) (_ :: proxy b) = forAll arbitrary $ \(s :: a) -> do
+    (eitherMigrate =<< migrateToB s) === Right s
+  where
+    migrateToB x = eitherMigrate x :: Either WalletError b
+
+migrateRoundtripProp
+    :: (Arbitrary a, Migrate a b, Migrate b a, Eq a, Show a, Typeable a, Typeable b)
+    => proxy a -> proxy b -> Spec
+migrateRoundtripProp proxyA proxyB =
+    prop ("Migrate " <> show (typeRep proxyA) <> " <-> " <> show (typeRep proxyB) <> " roundtrips") (migrateRoundtrip proxyA proxyB)
 
 aesonRoundtrip :: (Arbitrary a, ToJSON a, FromJSON a, Eq a, Show a) => proxy a -> Property
 aesonRoundtrip (_ :: proxy a) = forAll arbitrary $ \(s :: a) -> do
-    eitherDecode (encode (toJSON s)) === Right s
+    eitherDecode (encode s) === Right s
 
 aesonRoundtripProp
     :: (Arbitrary a, ToJSON a, FromJSON a, Eq a, Show a, Typeable a)
